+++ src/main/webapp/resources/js/pages/projects/samples/sample-ajax-params.js
@@ -13,12 +13,16 @@ function samplesAjaxParams() {
         var name = PAGE.ajaxParam.name;
         return typeof name === "string" ? name : "";
       },
+      organism: function() {
+        var organism = PAGE.ajaxParam.organism;
+        return typeof organism === "string" ? organism : "";
+      },
       minDate    : function () {
-        var minDate = PAGE.ajaxParam.minDate;
+        var minDate = PAGE.ajaxParam.date.min;
         return typeof minDate === "number" ? minDate : "";
       },
       endDate    : function () {
-        var endDate = PAGE.ajaxParam.endDate;
+        var endDate = PAGE.ajaxParam.date.end;
         return typeof endDate === "number" ? endDate : "";
       }
     }