+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes
 * [UI]: Fixed bug with selecting samples with associated projects. (0.13.3)
 * [UI]: Fixed bug when displaying samples names from associated projects when filtering by file. (0.13.3)
 * [UI]: Fixed a bug when selecting samples after filter by file with a large number of sample names. (0.13.4)
+* [UI]: Added filtering project samples by organism.
 * [API]: Added ability to upload sequencing runs as regular user as long as you have access to required projects.
 
 0.12.0 to 0.13.0